bitkeeper revision 1.1159.258.107 (4270fc89Gyl1f0UFludQff4kq09XvQ)
reboot.c, init.c, hypervisor.h, evtchn.h, tlbflush.h, pgtable-2level.h:
  sync w/ unstable.
system.h:
  Fix typo.
pgtable.h, fixup.c, irq.c, direct.c, timer_tsc.c, pci-dma.c, init.c, fault.c:
  Cleanup whitespace.
page.h:
  Turn __pte and __pgd into #define's like the original definitions.
  Add casts to unsigned long for phys_to_machine_mapping and
  machine_to_phys_mapping.
fixmap.h, Makefile:
  gnttab is not used in 2.0.
.del-gnttab.c~4461d10e936c37ab:
  Delete: linux-2.6.11-xen-sparse/arch/xen/kernel/gnttab.c
.del-gnttab.h~974eaf5d22418ecd:
  Delete: linux-2.6.11-xen-sparse/include/asm-xen/gnttab.h
setup.c:
  Add panic notification handler.
  Change phys_to_machine_mapping from unsigned long * to unsigned int *.
  Whitespace cleanups.
page.h:
  Change phys_to_machine_mapping from unsigned long * to unsigned int *.
process.c:
  UTSNAME -> system_utsname.release as in plain Linux.
  Disable preemption before reading smp_processor_id.
head.S:
  sync w/ unstable for maintenance -- backport SMP-guest support bits.
  Break __xen_guest definition up in several lines.
ldt.c, desc.h, common.c:
  Move queue flush into load_LDT().
common.c:
  g/c unused code.
Kconfig:
  Fix whitespace for microcode option.

+++ linux-2.6.11-xen-sparse/arch/xen/kernel/reboot.c
@@ -80,7 +80,7 @@ static void __do_suspend(void)
     extern void time_suspend(void);
     extern void time_resume(void);
     extern unsigned long max_pfn;
-    extern unsigned long *pfn_to_mfn_frame_list;
+    extern unsigned int *pfn_to_mfn_frame_list;
 
     suspend_record = (suspend_record_t *)__get_free_page(GFP_KERNEL);
     if ( suspend_record == NULL )
@@ -103,7 +103,8 @@ static void __do_suspend(void)
     HYPERVISOR_shared_info = (shared_info_t *)empty_zero_page;
     clear_fixmap(FIX_SHARED_INFO);
 
-    memcpy(&suspend_record->resume_info, &xen_start_info, sizeof(xen_start_info));
+    memcpy(&suspend_record->resume_info, &xen_start_info,
+	   sizeof(xen_start_info));
 
     HYPERVISOR_suspend(virt_to_machine(suspend_record) >> PAGE_SHIFT);
 
@@ -116,7 +117,8 @@ static void __do_suspend(void)
 
     shutting_down = -1; 
 
-    memcpy(&xen_start_info, &suspend_record->resume_info, sizeof(xen_start_info));
+    memcpy(&xen_start_info, &suspend_record->resume_info,
+	   sizeof(xen_start_info));
 
 #if LINUX_VERSION_CODE >= KERNEL_VERSION(2,6,0)
     set_fixmap_ma(FIX_SHARED_INFO, xen_start_info.shared_info);
